refactor: rename components into 'index'
