--- class-components/src/components/Button/types.ts
@@ -1,6 +1,0 @@
-export type ButtonProps = {
-  className: string;
-  type: "button" | "submit";
-  onClick?: () => void;
-  children?: React.ReactNode;
-};
