+++ class-components/src/components/ComponentWithError/index.tsx
