--- class-components/src/components/Input/types.ts
@@ -1,7 +1,0 @@
-export type InputProps = {
-  onClick?: () => void;
-  className: string;
-  name: string;
-  placeholder?: string;
-  defaultValue?: string;
-};
