--- class-components/src/components/Search/types.ts
@@ -1,5 +1,0 @@
-export type SearchProps = {
-  callback: Handler;
-};
-
-type Handler = (event: React.FormEvent<Element>) => Promise<void>;
